Refactor wifi into its own package
